Cleans authenticate and forms templates
This also introduces a new function, shibboleth_eppn(), which makes available
the `$_SERVER['redirect_eppn'] super-global to any template that needs it.
Ideally this would be done via an object, but that's a bit more than I have
time to take on right now.

See: https://phpmd.org/rules/controversial.txt on Superglobals