Only set brand secondary focus background if tabbing
Detect if user is tabbing and set focus pseudoclasses if they are.

https://medium.com/hackernoon/removing-that-ugly-focus-ring-and-keeping-it-too-6c8727fefcd2

This allows the hover navigation to not have undesired side effects if
a user clicks on one top menu item and then hovers over another.

A better longterm option might be to change the top level navigation to use buttons instead of a tags so we can remove the href which prevents the undesired effect but was going to lead to more aesthetic changes than I wanted to take on in this work.