Adds privacy notice JS locally (temporarily)
** Why are these changes being introduced:

We are preparing to roll out a common notice of our privacy policy
across many web platforms. At launch this will be loaded from a common
URL, but for now we are adding the prototype JS directly to this theme.

A future PR will update the enqueued JS to come from the common URL.

** Relevant ticket(s):

n/a

** How does this address that need:

Rather than needing to deploy this manually, it seems a good idea to
tackle the enqueuing and loading now. This commit is easy enough to
remove if I'm wrong here.

** Document any side effects to this change:

There is a section of "all-site JS" files that get enqueued without
conditionals. Now tht we have three of these, I've re-ordered them to
be sorted alphabetically.